separando logica de componente en login y register
